+++ src/query_search/core.clj
@@ -11,4 +11,4 @@
   [& args]
   (log "Запускаем сервер со следующими настройками:" (settings/get-settings))
   (println "Starting query-search server.")
-  (server/run-server handler {:port (settings/get-setting "port")})) ; Запускаем веб-сервер
+  (server/run-server handler {:port (settings/get-setting :port)})) ; Запускаем веб-сервер