+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 # Change Log
 
+## [Unreleased]
+## Added
+- Сервер для тестироваия ограничений по одновременным соединениям: testing/limiting-server
+- Сервер для тестирования API блогов: testing/fake-server
+## Changed
+- Concurrency теперь тестируется через limiting-server, а не crawler
+- Конфигурация разделена на default и dev
+
 ## 0.1.0 - 2017-04-07
 ### Added
 - Вся основная функциональность
 - Тесты
 - Документация
+
+[Unreleased]: https://github.com/ehpc/query-search-clojure/compare/v0.1.0...HEAD